refactor: BinaryContent import 경로 변경 및 불필요한 import 제거
