+++ discordit/src/main/java/com/sprint/mission/discodeit/dto/user/response/UserResponseDto.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.dto.user.response;
 
 import com.sprint.mission.discodeit.entity.base.User;
-import com.sprint.mission.discodeit.entity.content.BinaryContent;
+import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.enums.OnlineStatus;
 
 public record UserResponseDto(
